pyproject.toml
[project]
name = "dtach"
version = "0.30.2"
authors = [
    { name = "detachhead", email = "detachhead@users.noreply.github.com" },
]
description = "A Python tool to maintain a modular package architecture."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml~=6.0",
    "tomli>=1.2.2",
    "tomli-w~=1.0",
    "rich~=13.0",
    "prompt-toolkit~=3.0",
    "GitPython~=3.1",
    "networkx>=2.6,<4.0",
    "pydot>=2,<4",
    "stdlib-list>=0.10.0; python_version < '3.10'",
]
keywords = [
    'python',
    'module',
    'package',
    'guard',
    'enforcement',
    'boundary',
    'enforcer',
    'domain',
    'architecture',
]


[dependency-groups]
dev = [
    # Core deps (pinned)
    "rich==13.8.1",
    "prompt-toolkit==3.0.47",
    # Setup
    "pip==24.0",
    # Code Quality
    "basedpyright>=1.31.7",
    "ruff==0.11.5",
    # Build/Release
    "setuptools==69.5.1",
    "twine==5.1.1",
    "build==1.2.1",
    # Tests
    "pytest==8.2.2",
    "pytest-mock==3.14.0",
    "coverage==7.6.0",
    # Rust
    "maturin==1.7.1",
    "rustup>=1.28.2.1",
]
vscode = [
    "nodejs-wheel>=22.20.0",
    "nox>=2025.5.1",
    "PyHamcrest",
    "python-jsonrpc-server",
]


[project.urls]
Homepage = "https://github.com/detachhead/dtach"
Issues = "https://github.com/detachhead/dtach/issues"

[tool.ruff]
src = ["python"]
exclude = ["python/tests/example"]
target-version = "py38"
extend-exclude = [
    "pw",
]
lint.extend-select = ["I", "TC", "UP"]
lint.ignore = ["UP006", "UP007"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-decorators = ["dataclasses.dataclass"]
exempt-modules = ["typing", "typing_extensions"]

[tool.pyright]
include = ["python", "vscode"]
# https://github.com/DetachHead/basedpyright/issues/31
ignore = ["pw", "vscode/bundled", "vscode/.nox", "vscode/node_modules"]
exclude = ["pw", "vscode/bundled", "vscode/.nox", "vscode/node_modules"]
executionEnvironments = [{ root = "python" }, { root = "vscode" }]
pythonVersion = "3.8"


[tool.pytest.ini_options]
norecursedirs = ["python/tests/example", "vscode/bundled"]

[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "tach.extension"
features = ["pyo3/extension-module"]

[project.scripts]
tach = "tach.start:start"
